refactor : Issue #27 관련 수정사항 반영
