+++ src/feature/post/post.module.ts
@@ -1,10 +1,11 @@
+import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
 import { PostController } from './post.controller';
 import { PostService } from './post.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Post } from 'src/entity/post.entity';
+import { Post } from '../../entity/post.entity';
 import { PostLib } from './post.lib';
-import { HttpModule } from '@nestjs/axios';
 
 @Module({
   imports: [
