+++ src/feature/statistics/statistics.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { StatisticsService } from './statistics.service';
 import { StatisticsController } from './statistics.controller';
-import { PostModule } from '../post/post.app.module';
+import { PostModule } from '../post/post.module';
 
 @Module({
   imports: [PostModule],
